refactor: change path and names files
