+++ src/sass/components/_buttons.scss
@@ -35,7 +35,7 @@
         }
     
 
-        &_active {
+        &-active {
             display: block;
         }
     }
